iago: Redirect logs to user console in interactive installer
As per latest IAGO, the logs are redirected to display
when interactive installer is used. For this, the
terminal tty0 is opened.

We do not support the tty0 console. Instead ttyS0 is used,
which redirects the logs to user console.If we support
tty0 in the future, the ttyS0 should be updated to tty0.

Change-Id: I39fec24d6b7e2ed294dd1cdd3e0f8e42c546c1a2
Signed-off-by: Ravindranath D <ravindranathx.doddi@intel.com>